+++ packages/runtime-html/src/templating/controller.ts
@@ -46,7 +46,7 @@ import type { BindableDefinition } from '../bindable.js';
 import type { LifecycleHooksLookup } from './lifecycle-hooks.js';
 import type { INode, INodeSequence, IRenderLocation } from '../dom.js';
 import type { IViewFactory } from './view.js';
-import type { Instruction } from '../renderer.js';
+import type { IInstruction } from '../renderer.js';
 import type { IWatchDefinition, IWatcherCallback } from '../watch.js';
 import type { PartialCustomElementDefinition } from '../resources/custom-element.js';
 
@@ -130,11 +130,11 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   private readonly _rendering: IRendering;
 
   public readonly hooks: HooksDefinition;
+  public flags: LifecycleFlags = LifecycleFlags.none;
 
   public constructor(
     public container: IContainer,
     public readonly vmKind: ViewModelKind,
-    public flags: LifecycleFlags,
     public readonly definition: CustomElementDefinition | CustomAttributeDefinition | null,
     /**
      * The viewFactory. Only present for synthetic views.
@@ -153,8 +153,10 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
      */
     public host: HTMLElement | null,
   ) {
-    this.logger = null!;
-    this.debug = false;
+    if (__DEV__) {
+      this.logger = null!;
+      this.debug = false;
+    }
     this._rendering = container.root.get(IRendering);
     switch (vmKind) {
       case ViewModelKind.customAttribute:
@@ -196,7 +198,6 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     viewModel: C,
     host: HTMLElement,
     hydrationInst: IControllerElementHydrationInstruction | null,
-    flags: LifecycleFlags = LifecycleFlags.none,
     // Use this when `instance.constructor` is not a custom element type
     // to pass on the CustomElement definition
     definition: CustomElementDefinition | undefined = void 0,
@@ -210,7 +211,6 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     const controller = new Controller<C>(
       /* container      */ctn,
       /* vmKind         */ViewModelKind.customElement,
-      /* flags          */flags,
       /* definition     */definition,
       /* viewFactory    */null,
       /* viewModel      */viewModel as BindingContext<C>,
@@ -254,7 +254,6 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     ctn: IContainer,
     viewModel: C,
     host: HTMLElement,
-    flags: LifecycleFlags = LifecycleFlags.none,
     /**
      * The definition that will be used to hydrate the custom attribute view model
      *
@@ -271,7 +270,6 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     const controller = new Controller<C>(
       /* own ct         */ctn,
       /* vmKind         */ViewModelKind.customAttribute,
-      /* flags          */flags,
       /* definition     */definition,
       /* viewFactory    */null,
       /* viewModel      */viewModel as BindingContext<C>,
@@ -296,13 +294,11 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
    */
   public static $view(
     viewFactory: IViewFactory,
-    flags: LifecycleFlags = LifecycleFlags.none,
     parentController: ISyntheticView | ICustomElementController | ICustomAttributeController | undefined = void 0,
   ): ISyntheticView {
     const controller = new Controller(
       /* container      */viewFactory.container,
       /* vmKind         */ViewModelKind.synthetic,
-      /* flags          */flags,
       /* definition     */null,
       /* viewFactory    */viewFactory,
       /* viewModel      */null,
@@ -325,10 +321,12 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
      */
     hydrationContext: IHydrationContext | null,
   ): void {
-    this.logger = this.container.get(ILogger).root;
-    this.debug = this.logger.config.level <= LogLevel.debug;
-    if (this.debug) {
-      this.logger = this.logger.scopeTo(this.name);
+    if (__DEV__) {
+      this.logger = this.container.get(ILogger).root;
+      this.debug = this.logger.config.level <= LogLevel.debug;
+      if (__DEV__ && this.debug) {
+        this.logger = this.logger.scopeTo(this.name);
+      }
     }
 
     const container = this.container;
@@ -342,10 +340,10 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
       createWatchers(this, container, definition, instance);
     }
     createObservers(this, definition, flags, instance);
-    this._childrenObs = createChildrenObservers(this as Controller, definition, flags, instance);
+    this._childrenObs = createChildrenObservers(this as Controller, definition, instance);
 
     if (this.hooks.hasDefine) {
-      if (this.debug) { this.logger.trace(`invoking define() hook`); }
+      if (__DEV__ && this.debug) { this.logger.trace(`invoking define() hook`); }
       const result = instance.define(
         /* controller      */this as ICustomElementController,
         /* parentContainer */hydrationContext,
@@ -383,7 +381,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   /** @internal */
   public _hydrate(hydrationInst: IControllerElementHydrationInstruction | null): void {
     if (this.hooks.hasHydrating) {
-      if (this.debug) { this.logger!.trace(`invoking hydrating() hook`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`invoking hydrating() hook`); }
       (this.viewModel as BindingContext<C>).hydrating(this as ICustomElementController);
     }
 
@@ -420,7 +418,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     this.nodes = this._rendering.createNodes(compiledDef);
 
     if (this.hooks.hasHydrated) {
-      if (this.debug) { this.logger!.trace(`invoking hydrated() hook`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`invoking hydrated() hook`); }
       (this.viewModel as BindingContext<C>).hydrated(this as ICustomElementController);
     }
   }
@@ -428,7 +426,6 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   /** @internal */
   public _hydrateChildren(): void {
     this._rendering.render(
-      /* flags      */this.flags,
       /* controller */this as ICustomElementController,
       /* targets    */this.nodes!.findTargets(),
       /* definition */this._compiledDef!,
@@ -436,7 +433,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     );
 
     if (this.hooks.hasCreated) {
-      if (this.debug) { this.logger!.trace(`invoking created() hook`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`invoking created() hook`); }
       (this.viewModel as BindingContext<C>).created(this as ICustomElementController);
     }
   }
@@ -454,16 +451,16 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     this.lifecycleHooks = LifecycleHooks.resolve(this.container);
 
     if (this.hooks.hasCreated) {
-      if (this.debug) { this.logger!.trace(`invoking created() hook`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`invoking created() hook`); }
       (this.viewModel as BindingContext<C>).created(this as ICustomAttributeController);
     }
   }
 
+  /** @internal */
   private _hydrateSynthetic(): void {
     this._compiledDef = this._rendering.compile(this.viewFactory!.def!, this.container, null);
     this.isStrictBinding = this._compiledDef.isStrictBinding;
     this._rendering.render(
-      /* flags      */this.flags,
       /* controller */this as ISyntheticView,
       /* targets    */(this.nodes = this._rendering.createNodes(this._compiledDef)).findTargets(),
       /* definition */this._compiledDef,
@@ -509,7 +506,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     }
 
     this.parent = parent;
-    if (this.debug && !this._fullyNamed) {
+    if (__DEV__ && this.debug && !this._fullyNamed) {
       this._fullyNamed = true;
       (this.logger ??= this.container.get(ILogger).root.scopeTo(this.name)).trace(`activate()`);
     }
@@ -549,7 +546,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     this._enterActivating();
 
     if (this.hooks.hasBinding) {
-      if (this.debug) { this.logger!.trace(`binding()`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`binding()`); }
 
       const ret = this.viewModel!.binding(this.$initiator, this.parent, this.$flags);
       if (ret instanceof Promise) {
@@ -568,7 +565,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   }
 
   private bind(): void {
-    if (this.debug) { this.logger!.trace(`bind()`); }
+    if (__DEV__ && this.debug) { this.logger!.trace(`bind()`); }
 
     let i = 0;
     let ii = this._childrenObs.length;
@@ -595,7 +592,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     }
 
     if (this.hooks.hasBound) {
-      if (this.debug) { this.logger!.trace(`bound()`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`bound()`); }
 
       ret = this.viewModel!.bound(this.$initiator, this.parent, this.$flags);
       if (ret instanceof Promise) {
@@ -633,7 +630,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   }
 
   private _attach(): void {
-    if (this.debug) { this.logger!.trace(`attach()`); }
+    if (__DEV__ && this.debug) { this.logger!.trace(`attach()`); }
 
     if (this.hostController !== null) {
       switch (this.mountTarget) {
@@ -666,7 +663,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     }
 
     if (this.hooks.hasAttaching) {
-      if (this.debug) { this.logger!.trace(`attaching()`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`attaching()`); }
 
       const ret = this.viewModel!.attaching(this.$initiator, this.parent, this.$flags);
       if (ret instanceof Promise) {
@@ -716,7 +713,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
           throw new Error(`AUR0505:${this.name} ${stringifyState(this.state)}`);
     }
 
-    if (this.debug) { this.logger!.trace(`deactivate()`); }
+    if (__DEV__ && this.debug) { this.logger!.trace(`deactivate()`); }
 
     this.$initiator = initiator;
     this.$flags = flags;
@@ -743,7 +740,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
     }
 
     if (this.hooks.hasDetaching) {
-      if (this.debug) { this.logger!.trace(`detaching()`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`detaching()`); }
 
       const ret = this.viewModel!.detaching(this.$initiator, this.parent, this.$flags);
       if (ret instanceof Promise) {
@@ -805,7 +802,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   }
 
   private unbind(): void {
-    if (this.debug) { this.logger!.trace(`unbind()`); }
+    if (__DEV__ && this.debug) { this.logger!.trace(`unbind()`); }
 
     const flags = this.$flags | LifecycleFlags.fromUnbind;
     let i = 0;
@@ -895,7 +892,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   private _leaveActivating(): void {
     if (--this._activatingStack === 0) {
       if (this.hooks.hasAttached) {
-        if (this.debug) { this.logger!.trace(`attached()`); }
+        if (__DEV__ && this.debug) { this.logger!.trace(`attached()`); }
 
         _retPromise = this.viewModel!.attached!(this.$initiator, this.$flags);
         if (_retPromise instanceof Promise) {
@@ -932,7 +929,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   private _leaveDetaching(): void {
     if (--this._detachingStack === 0) {
       // Note: this controller is the initiator (detach is only ever called on the initiator)
-      if (this.debug) { this.logger!.trace(`detach()`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`detach()`); }
 
       this._enterUnbinding();
       this.removeNodes();
@@ -974,7 +971,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   }
   private _leaveUnbinding(): void {
     if (--this._unbindingStack === 0) {
-      if (this.debug) { this.logger!.trace(`unbind()`); }
+      if (__DEV__ && this.debug) { this.logger!.trace(`unbind()`); }
 
       let cur = this.$initiator.head as Controller | null;
       let next: Controller | null = null;
@@ -1063,7 +1060,7 @@ export class Controller<C extends IViewModel = IViewModel> implements IControlle
   }
 
   public dispose(): void {
-    if (this.debug) { this.logger!.trace(`dispose()`); }
+    if (__DEV__ && this.debug) { this.logger!.trace(`dispose()`); }
 
     if ((this.state & State.disposed) === State.disposed) {
       return;
@@ -1167,8 +1164,6 @@ function createObservers(
 function createChildrenObservers(
   controller: Controller,
   definition: CustomElementDefinition,
-  // deepscan-disable-next-line
-  _flags: LifecycleFlags,
   instance: object,
 ): ChildrenObserver[] {
   const childrenObservers = definition.childrenObservers;
@@ -1742,11 +1737,10 @@ export interface ICustomElementViewModel extends IViewModel, IActivationHooks<IH
 export interface ICustomAttributeViewModel extends IViewModel, IActivationHooks<IHydratedController> {
   readonly $controller?: ICustomAttributeController<this>;
   link?(
-    flags: LifecycleFlags,
     controller: IHydratableController,
     childController: ICustomAttributeController,
     target: INode,
-    instruction: Instruction,
+    instruction: IInstruction,
   ): void;
   created?(
     controller: ICustomAttributeController<this>,