+++ packages/runtime/rollup.config.js
@@ -23,7 +23,8 @@ const terserPluginCfg = terser({
   },
   mangle: {
     properties: {
-      regex: /^_/
+      regex: /^_/,
+      reserved: ['__esModule']
     }
   },
   format: {
@@ -45,19 +46,21 @@ export default {
       file: `dist/esm/index.js`,
       format: 'es',
       sourcemap: true,
-      plugins: [terserPluginCfg]
+      plugins: [terserPluginCfg],
     },
     {
       file: `dist/cjs/index.dev.js`,
       format: 'cjs',
       sourcemap: true,
       esModule: true,
+      externalLiveBindings: false,
     },
     {
       file: `dist/cjs/index.js`,
       format: 'cjs',
       sourcemap: true,
-      plugins: [terserPluginCfg]
+      externalLiveBindings: false,
+      plugins: [terserPluginCfg],
     },
   ],
   plugins: [