+++ packages/validation-html/CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.0-alpha.15"></a>
+# 2.0.0-alpha.15 (2021-08-01)
+
+**Note:** Version bump only for package @aurelia/validation-html
+
 <a name="2.0.0-alpha.14"></a>
 # 2.0.0-alpha.14 (2021-07-25)